% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helpers.R
\name{build_design_and_contrast}
\alias{build_design_and_contrast}
\title{Build limma design matrix and contrast}
\usage{
build_design_and_contrast(
  metadata,
  condition_column,
  group1_condition,
  group2_condition,
  covariates = NULL,
  design_formula = NULL,
  contrast_string = NULL
)
}
\description{
Build limma design matrix and contrast
}
\keyword{internal}
