% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_packages}
\alias{check_packages}
\title{Check if a series of Packages is installed}
\usage{
check_packages(required_pkgs)
}
\arguments{
\item{required_pkgs}{vector of names of the required packages}
}
\value{
Prints uninstalled packages
}
\description{
Check if a series of Packages is installed
}
