% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution.R
\name{create_immune_boxplots}
\alias{create_immune_boxplots}
\title{Create Immune Cell Boxplots}
\usage{
create_immune_boxplots(
  immune_scores,
  metadata,
  condition_column,
  output_dir,
  experiment_name,
  scale = TRUE
)
}
\arguments{
\item{immune_scores}{named list with Immune scores matrices per tool}

\item{metadata}{Sample metadata}

\item{condition_column}{Condition column name}

\item{output_dir}{Output directory}

\item{experiment_name}{Experiment name}

\item{scale}{whether or not to scale the scores (default: TRUE)}
}
\value{
ggplot object
}
\description{
Creates boxplots comparing immune cell abundances between conditions.
}
