% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_analysis.R
\name{get_hub_genes}
\alias{get_hub_genes}
\title{Get Hub Genes from WGCNA Modules}
\usage{
get_hub_genes(datExpr, moduleColors, n_hub_genes = 10)
}
\arguments{
\item{datExpr}{Expression data used in WGCNA}

\item{moduleColors}{Module assignments}

\item{n_hub_genes}{Number of hub genes to return per module (default: 10)}
}
\value{
Data frame with hub genes for each module
}
\description{
Identifies hub genes (highly connected genes) within each module.
}
