% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_palette}
\alias{get_palette}
\title{Creating palette}
\usage{
get_palette(n_conditions, val_conditions)
}
\arguments{
\item{n_conditions}{Number of conditions to colour}

\item{val_conditions}{Values of conditions to colour}
}
\value{
a named list with colours for each condition
}
\description{
Creating palette
}
