% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution.R
\name{harmonize_deconvolution}
\alias{harmonize_deconvolution}
\title{Harmonize Deconvolution outputs}
\usage{
harmonize_deconvolution(deconvo_output, tool_name)
}
\arguments{
\item{deconvo_output}{the output of the deconvolution method}

\item{tool_name}{the tool name (CIBERSORTx, xCell, EPIC-BRef, EPIC-TRef, quanTIseq, MCPC)}
}
\value{
an harmonized data.frame
}
\description{
For each tool it provides common labels and aggregates fine types into coarse ones whenever possible
}
