% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_analysis.R
\name{plot_dendrogram_and_modules}
\alias{plot_dendrogram_and_modules}
\title{Plot Dendrogram and Modules}
\usage{
plot_dendrogram_and_modules(
  geneTree,
  moduleColors,
  output_dir,
  experiment_name
)
}
\arguments{
\item{geneTree}{gene Tree}

\item{moduleColors}{colors for modules}

\item{output_dir}{path of the output directory}

\item{experiment_name}{Name of the experiment}
}
\value{
a plot
}
\description{
Plot Dendrogram and Modules
}
\keyword{internal}
