% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_gene_expression}
\alias{plot_gene_expression}
\title{Plot of Gene Expression for a selected set of genes (log(TPM+1) is used)}
\usage{
plot_gene_expression(
  results,
  gene_vector,
  output_dir,
  group_by = "cell_line",
  facet_by = "condition",
  remove_samples = NULL,
  species = "human",
  remove_id = TRUE,
  stat_test = "wilcox.test",
  p_correction = "fdr",
  col_cluster = TRUE,
  row_cluster = TRUE,
  plot_title = NULL,
  ext = "pdf",
  file_name = "Expression_heatmap_of_selected_genes",
  ...
)
}
\arguments{
\item{results}{the object obtained by running run_complete_pipeline}

\item{gene_vector}{a vector of gene symbols or ENSEMBL IDs (or mixture)}

\item{output_dir}{the path where to save the plot}

\item{group_by}{name of the column whose values will correspond to a single barplot in the plot}

\item{facet_by}{name of the column whose values will correspond to a single plot}

\item{remove_samples}{names of the samples to remove}

\item{species}{either "human" or "mouse" (default: "human")}

\item{remove_id}{whether to remove gene ID in the heatmap (default: TRUE)}

\item{stat_test}{what test to use in the boxplot (default: "wilcox.test")}

\item{p_correction}{what p-val correction to use in the boxplot (default: "fdr")}

\item{col_cluster}{bool to cluster columns (default: TRUE)}

\item{row_cluster}{bool to cluster rows (default: TRUE)}

\item{plot_title}{name to assign to the plot (default: NULL)}

\item{ext}{file extension (default: "pdf")}

\item{file_name}{(default: "Expression_heatmap_of_selected_genes")}

\item{...}{extra parameters for create_expression_heatmap()}
}
\value{
prints the plot and saves it in the specified directory
}
\description{
Plot of Gene Expression for a selected set of genes (log(TPM+1) is used)
}
