% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_analysis.R
\name{plot_module_eigengenes}
\alias{plot_module_eigengenes}
\title{Plot Module Eigengenes}
\usage{
plot_module_eigengenes(MEs, trait_data, output_dir, experiment_name)
}
\arguments{
\item{MEs}{Module eigengenes}

\item{trait_data}{Trait data}

\item{output_dir}{Output directory}

\item{experiment_name}{Experiment name}
}
\value{
None (creates plot)
}
\description{
Creates plots showing module eigengene expression patterns.
}
\keyword{internal}
