% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_analysis.R
\name{plot_module_trait_heatmap}
\alias{plot_module_trait_heatmap}
\title{Plot Module-Trait Heatmap}
\usage{
plot_module_trait_heatmap(
  moduleTraitCor,
  moduleTraitP,
  output_dir,
  experiment_name
)
}
\arguments{
\item{moduleTraitCor}{Module-trait correlation matrix}

\item{moduleTraitP}{Module-trait p-value matrix}

\item{output_dir}{Output directory}

\item{experiment_name}{Experiment name}
}
\value{
None (creates plot)
}
\description{
Creates a heatmap showing correlations between modules and traits.
}
\keyword{internal}
