% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_analysis.R
\name{plot_tf_volcano}
\alias{plot_tf_volcano}
\title{Create Volcano Plot for TF activity}
\usage{
plot_tf_volcano(
  diff_tf_results,
  output_dir,
  experiment_name,
  fdr_threshold = 0.05,
  lfc_threshold = 1,
  n_labels = 20
)
}
\arguments{
\item{diff_tf_results}{the results of TF differential activity analysis}

\item{output_dir}{path to the output directory}

\item{experiment_name}{name of the experiment}

\item{fdr_threshold}{FDR threshold (default: 0.05)}

\item{lfc_threshold}{log-Fold-Change threshold (default: 1)}

\item{n_labels}{number of labels to be shown (default: 20)}
}
\value{
a volcano plot
}
\description{
Create Volcano Plot for TF activity
}
