% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PvalCalc}
\alias{PvalCalc}
\title{P-value Calculator for the fn plot_gene_expression()}
\usage{
PvalCalc(data, facet.col, y.col, x.col, stat_test, p_correction = "none")
}
\arguments{
\item{data}{a dataframe with a column containing the gene expr values and more}

\item{facet.col}{name of the column according to which group}

\item{y.col}{name of the column with gene data}

\item{x.col}{name of the column of inside grouping}

\item{stat_test}{the kind of stat_test to perform}

\item{p_correction}{the type of p_correction to implement (default: "none")}
}
\value{
df with the p-vals for the various comparisons
}
\description{
P-value Calculator for the fn plot_gene_expression()
}
