% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{retrieve_pathway}
\alias{retrieve_pathway}
\title{Retrive Custom Pathways}
\usage{
retrieve_pathway(file.path, species, matrix_row_names = NULL)
}
\arguments{
\item{file.path}{the pathway to an excel file with each sheet named
after the specific pathway and inside a table with at least one of the
following cols (named like that): "NCBI", "ENSEMBL", and/or "Symbol"}

\item{species}{either "human" or "mouse"}

\item{matrix_row_names}{the row names of the RNA-Seq data matrix (if available)}
}
\value{
a named list with the pathways and the respective genes
}
\description{
Retrive Custom Pathways
}
