% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution.R
\name{run_tools}
\alias{run_tools}
\title{Run all human Deconvolution Tools}
\usage{
run_tools(exprMatrix, tools = "all", output_dir = NULL, experiment_name = NULL)
}
\arguments{
\item{exprMatrix}{a matrix with raw TPM (not log-transformed), and gene symbols as rownames}

\item{tools}{a vector of tool names to be used (all, xCell, EPIC-TRef, EPIC-BRef, quanTIseq, MCPC)}

\item{output_dir}{the path where to save the results}

\item{experiment_name}{the name to give to the results}
}
\value{
a named list with the deconvolution output per tool
}
\description{
Performs Immune Deconvolution using all the listed tools
}
