% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{save_gene_lists}
\alias{save_gene_lists}
\title{Save Gene Lists}
\usage{
save_gene_lists(
  fit,
  experiment_name,
  output_dir,
  fdr_threshold,
  lfc_threshold,
  fit_label = "efit"
)
}
\arguments{
\item{fit}{fit object}

\item{experiment_name}{Name of the experiment}

\item{output_dir}{path to the output directory}

\item{fdr_threshold}{FDR threshold}

\item{lfc_threshold}{log-Fold-Change threshold}

\item{fit_label}{Label for the fit (default: "efit")}
}
\description{
Save Gene Lists
}
\keyword{internal}
