% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_analysis.R
\name{save_wgcna_results}
\alias{save_wgcna_results}
\title{Save WGCNA Results}
\usage{
save_wgcna_results(
  moduleColors,
  MEs,
  module_trait_results,
  gene_module_analysis,
  output_dir,
  experiment_name
)
}
\arguments{
\item{moduleColors}{Module color assignments}

\item{MEs}{Module eigengenes}

\item{module_trait_results}{Module-trait correlation results}

\item{gene_module_analysis}{Gene significance analysis results}

\item{output_dir}{Output directory}

\item{experiment_name}{Experiment name}
}
\value{
None (invisible)
}
\description{
Saves all WGCNA results to files.
}
\keyword{internal}
