% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{save_gene_lists}
\alias{save_gene_lists}
\title{Save Gene Lists}
\usage{
save_gene_lists(
  fit,
  experiment_name,
  output_dir,
  fdr_threshold,
  lfc_threshold,
  fit_label = "efit"
)
}
\description{
Save Gene Lists
}
\keyword{internal}
