% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{summarize_de_results}
\alias{summarize_de_results}
\title{Summarize Differential Expression Results}
\usage{
summarize_de_results(
  fit,
  fdr_threshold = 0.1,
  lfc_threshold = 1,
  fit_label = "efit"
)
}
\description{
Summarize Differential Expression Results
}
\keyword{internal}
